--- .github/workflows/test-release.yml
@@ -1,26 +1,0 @@
-name: Test Release
-
-on:
-  push:
-
-concurrency: ${{ github.workflow }}-${{ github.ref }}
-
-jobs:
-  release:
-    name: Test Release
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-        with:
-          # This makes Actions fetch all Git history so that Changesets can generate changelogs with the correct commits
-          fetch-depth: 0
-          # This makes the PR pushed to use GITHUB_TOKEN and trigger the checks
-          persist-credentials: false
-
-  publish-cli-assets:
-    name: Publish CLI assets
-    needs: [release]
-    uses: ./.github/workflows/release-cli-assets.yml
-    with:
-      commitSha: ${{ github.sha }}
-    secrets: inherit
